requirements.txt
asgiref             
Django              
djangorestframework 
pip                 
pytz                
setuptools          
sqlparse            
typing-extensions   
django-cors-headers 
django-mysql        
mysqlclient         
pika
django-filter
pillow                
